Extract config tests and tidy exports
