+++ docs/configuration-roadmap.md
@@ -72,7 +72,8 @@ expanded with specifics for the configuration design.
     or unsupported features like** `incremental=True`**.**
 
 - [x] Define the `FemtoFilter` trait and implement common filter types (e.g.,
-  `LevelFilter`, `NameFilter`), with builder API integration for filters.
+  `LevelFilter`, `NameFilter`), with builder API integration for
+  filters.[^pr198-filters]
 
 - [ ] Expand test coverage and start benchmarking for the configuration system
   and `basicConfig`/`dictConfig` compatibility.
@@ -96,6 +97,8 @@ expanded with specifics for the configuration design.
 - [ ] Investigate and implement more sophisticated dynamic reconfiguration
   capabilities for handlers and filters at runtime via the builder API (as a
   V1.1 or V2 feature).
+
+[^pr198-filters]: completed in PR #198 on 6 September 2025.
 
 - [ ] Provide a `tracing_subscriber::Layer` so `femtologging` handlers can
   process `tracing` spans and events.
