+++ femtologging/__init__.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 # Import the Rust extension packaged under this module's namespace first
 # to keep imports at the top for linters.
-from . import _femtologging_rs as rust  # type: ignore[attr-defined]
+from . import _femtologging_rs as rust
 from .overflow_policy import OverflowPolicy
 from .config import dictConfig
 import logging
@@ -24,23 +24,23 @@ if _typing.TYPE_CHECKING:
         NameFilterBuilder as NameFilterBuilder,
     )
 
-hello = rust.hello  # type: ignore[attr-defined]
-FemtoLogger = rust.FemtoLogger  # type: ignore[attr-defined]
-get_logger = rust.get_logger  # type: ignore[attr-defined]
-reset_manager = rust.reset_manager_py  # type: ignore[attr-defined]
-FemtoHandler = rust.FemtoHandler  # type: ignore[attr-defined]
-FemtoStreamHandler = rust.FemtoStreamHandler  # type: ignore[attr-defined]
-FemtoFileHandler = rust.FemtoFileHandler  # type: ignore[attr-defined]
-StreamHandlerBuilder = rust.StreamHandlerBuilder  # type: ignore[attr-defined]
-FileHandlerBuilder = rust.FileHandlerBuilder  # type: ignore[attr-defined]
-ConfigBuilder = rust.ConfigBuilder  # type: ignore[attr-defined]
-LoggerConfigBuilder = rust.LoggerConfigBuilder  # type: ignore[attr-defined]
-FormatterBuilder = rust.FormatterBuilder  # type: ignore[attr-defined]
-LevelFilterBuilder = rust.LevelFilterBuilder  # type: ignore[attr-defined]
-NameFilterBuilder = rust.NameFilterBuilder  # type: ignore[attr-defined]
-FilterBuildError = rust.FilterBuildError  # type: ignore[attr-defined]
-HandlerConfigError = rust.HandlerConfigError  # type: ignore[attr-defined]
-HandlerIOError = rust.HandlerIOError  # type: ignore[attr-defined]
+hello = rust.hello
+FemtoLogger = rust.FemtoLogger
+get_logger = rust.get_logger
+reset_manager = rust.reset_manager_py
+FemtoHandler = rust.FemtoHandler
+FemtoStreamHandler = rust.FemtoStreamHandler
+FemtoFileHandler = rust.FemtoFileHandler
+StreamHandlerBuilder = rust.StreamHandlerBuilder
+FileHandlerBuilder = rust.FileHandlerBuilder
+ConfigBuilder = rust.ConfigBuilder
+LoggerConfigBuilder = rust.LoggerConfigBuilder
+FormatterBuilder = rust.FormatterBuilder
+LevelFilterBuilder = rust.LevelFilterBuilder
+NameFilterBuilder = rust.NameFilterBuilder
+FilterBuildError = rust.FilterBuildError
+HandlerConfigError = rust.HandlerConfigError
+HandlerIOError = rust.HandlerIOError
 
 
 @dataclass
